Address migration review feedback
