+++ server/src/main/java/com/github/klboke/kkrepo/server/cargo/CargoProxyService.java
@@ -15,7 +15,6 @@ import com.github.klboke.kkrepo.protocol.cargo.CargoPath;
 import com.github.klboke.kkrepo.protocol.cargo.CargoVersions;
 import com.github.klboke.kkrepo.server.cache.AssetMetadataCache;
 import com.github.klboke.kkrepo.server.cache.CachedAssetMetadata;
-import com.github.klboke.kkrepo.server.cargo.CargoPublishPayload.CargoPackageMetadata;
 import com.github.klboke.kkrepo.server.maven.BlobStorageRegistry;
 import com.github.klboke.kkrepo.server.maven.HttpRemoteFetcher;
 import com.github.klboke.kkrepo.server.maven.MavenResponse;
